* Added bandwidth limits and deadlines to the QoS_Transport_Settings class   * set_send_bandwidth_limit can be used to limit the amount sent according to what has been previously sent in the past 10s   * set_total_bandwidth_limit can be used to limit the amount sent according to what has been previously received in the past 10s   * set_deadline can be used to reject received messages that are past a latency deadline in seconds * Added processing logic for deadline and bandwidth decisions within all supported transports * Added tutorials for setting send and total bandwidth limits inside of tutorials/advanced * Changed bandwidth monitors to keep 64 bit counters instead of 32 bit * Updated to version 1.1.10. Version 1.2.0 will be tagged once we have tested the new features sufficiently.
git-svn-id: http://madara.googlecode.com/svn/trunk@935 2850746a-f1c8-11dd-b507-8b4ecd5210be

+++ include/madara/transport/Transport_Context.cpp
@@ -3,7 +3,7 @@
 #include "madara/utility/Log_Macros.h"
 
 Madara::Transport::Transport_Context::Transport_Context (int64_t operation,
-        uint32_t receive_bandwidth, uint32_t send_bandwidth,
+        uint64_t receive_bandwidth, uint64_t send_bandwidth,
         uint64_t message_time, uint64_t current_time)
   : operation_ (operation),
     receive_bandwidth_ (receive_bandwidth),
